pip_backend/requirements.txt
asgiref==3.5.2

certifi==2022.9.14
cloudinary==1.30.0
dj-database-url==1.0.0
Django
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
mongoengine==0.24.2
Pillow==9.2.0
PyJWT==2.5.0
pymongo==3.12.3
python-decouple==3.6
pytz==2022.2.1
six==1.16.0
sqlparse==0.2.4
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0